Prep-file generation now uses sample_name, not well_description.
Sample_name is now a required column.
Well_description is no longer a required column. However a warning message will be issued if well_description is not present, as it is customarily present.
If well_description is present as 'description', it will be renamed to 'well_description' and a warning will be issued as was done previously.
Contents of 'filtered' directory renamed to 'filtered_seqeuences' directory and unittest fixed to run as intended.
Several sample fastq and trimmed/filtered files had their names changed slightly to deobfuscate when a path is built using sample_name, sample_id, and whatnot.