Added second Stats.json file
Added second Stats.json file containing the same set of IndexSequences
across all lanes. This allows us to test the summing of the appropriate
DataFrames properly.
Added second unittest for extract_stats_metadata() and sum_lanes() to
test extraction and summing of values across multiple lanes, in reverse
of the order of appearance in the original file.