User-defined well_description now working correctly.
Previously, make_sample_sheet() would ignore 'Well_description' if it was defined in the input table and instead would generate the column from sample-name values.
This fix will allow 'Well_description' values to reflect as intended when using make_sample_sheet() and accept 'description' and 'well_description' as alternatives in the input table as well.
Code-review confirmed that 'Well_description' is no longer used to obtain the sample-name in metapool.