Updates to replicate functionality
Updates to replicate functionality, based on discussions with Rodolfo
and Antonio. Sample-sheets and prep-files are now broken up only by
plate and not by (project, plate). It is now assumed that all projects
w/in a sample-sheet or pre-prep file will either all contain replicates,
or all will not contain replicates. This is a lab convention, along
w/plate-level replication.