Replicate updates (#134)
* Updates to replicate functionality

Updates to replicate functionality, based on discussions with Rodolfo
and Antonio. Sample-sheets and prep-files are now broken up only by
plate and not by (project, plate). It is now assumed that all projects
w/in a sample-sheet or pre-prep file will either all contain replicates,
or all will not contain replicates. This is a lab convention, along
w/plate-level replication.

* Added new files

* Piggybacking read_plate_map_csv() fix

This fix to read_plate_map_csv() preserves legacy tests and changes
absolute minimum for this legacy function. Post fix, the 'Sample' column
will always be interpreted as a string. This prevents Errors when
Samples are all numbers e.g. '363167879'.

* test pinning scikit-bio to 0.5.8 to resolve compilation issue

* Updates based on added tests to qp-klp.