prep columns defined in sample-sheet classes
Sample-sheet child classes now contain a list of the columns required
for a prep-info file using that variant of sample-sheet as input.
Unknown columns in sample-sheets will not automatically pass through to
the prep-info file, but they can be added to the correct lists and they
will be. This is more desirable than transparent pass-thru.
prep-info-file generation cleaned up.