seqpro updated to pass through columns (#179)
* seqpro updated to pass through columns

seqpro updated to pass through columns specific to a particular
KLSampleSheet child-type.

* Added new sample-sheet

* Small clean-up of legacy code

* prep columns defined in sample-sheet classes

Sample-sheet child classes now contain a list of the columns required
for a prep-info file using that variant of sample-sheet as input.
Unknown columns in sample-sheets will not automatically pass through to
the prep-info file, but they can be added to the correct lists and they
will be. This is more desirable than transparent pass-thru.
prep-info-file generation cleaned up.