Removed sanity-check to support plate-level rep
Removed sanity-check that raises an error if run_counts() finds a
fastq.gz file that isn't mentioned in the sample-sheet. This assumption
does not hold during plate-level replication where new sample-sheets are
created for each replicate. In this case, each new sample-sheet is by
definition going to contain only a subset of the sample_ids found in the
fastq.gz directory.