Cleaned up test_data input and output for notebook templates (#229)
Generated new output files for iSeqnormpool because we changed default behavior to normalized pooling on 'Raw Reads' and the previous output file was not reflecting this. Separated the test_ouput files for the matrix notebook and the legacy notebook (plate map based) because they were overwriting the same output file with small sample_name discrepancies that resulted in unstable changes to files in repository depending on which notebook was being run last before committing changes. Fixed a small issue with katharoseq barcodes being present in matrix tube template when we instead are trying to default to a metagenomic template with no katharoseq samples.