Added function to parse a sample-sheet into dataframes.
Extended _parse_header() to parse all sections of a sample-sheet into
dataframes and return them to the caller. This will make it easier and
more robust to modify values like 'Lane' on-demand. This is part of an
effort to potentially replace the third-party sample_sheet module and
parsing method. Lane is a field that is currently read-only once read
into the base SampleSheet() object. A robust method to change the values
of an existing column or add a column if it's not present is needed.