Updated _parse_header()
Updated _parse_header() into a small DFSheet() class that can parse a
sample-sheet into dataframes w/out any verification. The user can
manually edit the dataframes and there is a method to write the
DataFrames back out to the same csv file. The _parse_header() function
is now a method of the DFSheet() class.

The goal is for this functionality to be merged w/KLSampleSheet and
remove the dependency on the third party sample_sheet object, which
doesn't allow manipulation of the [Data] segment. It can also replace
the legacy _parse() method, which relies on the csv and sample_sheet
packages and can be hard to modify due to its line-by-line processing.