Removed legacy check that no longer applies.
Removed check from sample-sheet's _parse() function. This function
assumes that any trailing ',' in the header directly below a [Data]
section header is a column w/out a name.

This only holds true when the [Data] section has the largest number of
columns in the sheet, which has historically been true. However, it is
now possible for [Bioinformatics] to have more columns than [Data] w/the
new tellseq sample-sheets and hence trigger this error incorrectly.

Removed check as validation downstream of _parse() will catch a blank
column name. There's no need to do it in _parse().