Added code for fix. (#280)
* Added code for fix.
Updated legacy test to conform to current assumptions.
 1) sample-sheets aren't expected to contain samples from multiple lanes since SPP runs are on a per-lane basis.
 2) sample-names while technically are unique to a project need to be unique per sample-sheet to accomodate legacy seqpro df-munging code.

* Broken tests reviewed and updated as needed.

* Updated existing tests

Updated some existing tests that were affected by recent changes.
Removed legacy atropos_qc directories as they no longer reflect current code or operation.

* Working replacement.

Tests aren't merged in.

* Updated test

seqpro counts test updated.
Test only fakes files for two of the samples of an actual sample-sheet.
Valid warnings were being generated for the non-faked samples.
A new test sample-sheet was created containing only the two samples so that stderr did not need to be suppressed in the test.

* Added tests for _map_files_to_sample_ids().

* minor updates

* All new tests added

* All tests working

* seqpro will now write failed samples to disk.

Sample-ids that could not be mapped to a pair of fastq files will be written to a file in the output dir named 'unmatched.json'.

* flake8

* Removed junk file

* Updates to handle SPP replicates

* Tests updated

* Updated fillna

---------

Co-authored-by: qiita_t <qiita_t@qiita-test-2.ucsd.edu>